move crypto object test cases to new file & check sentinel while decrypt returned
